# pyparsers/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/cars/dongchedi/page/{page}": {
            "get": {
                "summary": "Get Dongchedi Cars By Page",
                "description": "Получает данные о машинах с dongchedi для конкретной страницы\n\nArgs:\n    page: Номер страницы (начиная с 1)",
                "operationId": "get_dongchedi_cars_by_page_cars_dongchedi_page__page__get",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Page"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cars/dongchedi/incremental": {
            "post": {
                "summary": "Get Dongchedi Incremental Cars",
                "description": "Получает только новые машины с dongchedi до первого совпадения с существующими.\n\nArgs:\n    existing_cars: Список существующих машин с полями car_id/sku_id/link",
                "operationId": "get_dongchedi_incremental_cars_cars_dongchedi_incremental_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "additionalProperties": true,
                                    "type": "object"
                                },
                                "type": "array",
                                "title": "Existing Cars"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cars/dongchedi/car/{car_id}": {
            "get": {
                "summary": "Get Dongchedi Car Detail",
                "description": "Получает детальную информацию о конкретной машине с dongchedi по ID\n\nArgs:\n    car_id: ID машины в dongchedi\n    \nReturns:\n    Dict с информацией о машине, включая статус продажи",
                "operationId": "get_dongchedi_car_detail_cars_dongchedi_car__car_id__get",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Car Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}